+++ packages/ricos-types/web/src/pluginMenuTypes.ts
@@ -1,0 +1,3 @@
+export const PLUGIN_MENU_MODAL_ID = 'pluginMenu';
+
+export const PLUGIN_MENU_HORIZONTAL_MODAL_ID = 'pluginMenuHorizontal';