+++ packages/toolbars-ui/web/src/components/AddPluginMenu/consts.ts
@@ -1,1 +1,0 @@
-export const PLUGIN_MENU_MODAL_ID = 'pluginMenu';