+++ packages/toolbars-ui/web/src/components/AddPluginMenu/index.ts
@@ -1,2 +1,1 @@
 export { default as AddPluginMenu } from './AddPluginMenu';
-export { PLUGIN_MENU_MODAL_ID } from './consts';