+++ packages/toolbars-v3/web/src/modals/FloatingPluginMenu/AddPluginMenuHorizontal.tsx
@@ -1,14 +1,14 @@
 import React, { useContext } from 'react';
 import { EditorContext } from 'ricos-context';
-import type { PluginAddButton, RicosEditorPlugins } from 'ricos-types';
+import type { PluginAddButton, PluginAddButtons } from 'ricos-types';
 import InsertPluginToolbar from '../../components/InsertPluginToolbar/InsertPluginToolbar';
 import styles from './styles/floating-add-plugin-menu.scss';
 interface Props {
   referenceElement?: React.RefObject<HTMLElement>;
-  plugins: RicosEditorPlugins;
+  buttons: PluginAddButtons;
 }
 
-const AddPluginMenuHorizontal: React.FC<Props> = ({ referenceElement, plugins }) => {
+const AddPluginMenuHorizontal: React.FC<Props> = ({ referenceElement, buttons }) => {
   const { getEditorCommands } = useContext(EditorContext);
 
   const onButtonClick = (addButton: PluginAddButton) => {
@@ -18,7 +18,7 @@ const AddPluginMenuHorizontal: React.FC<Props> = ({ referenceElement, plugins })
 
   return (
     <div className={styles.addPluginMenu_horizontal_wrapper} data-hook="addPluginMenuHorizontal">
-      <InsertPluginToolbar buttons={plugins.getAddButtons()} onButtonClick={onButtonClick} />
+      <InsertPluginToolbar buttons={buttons} onButtonClick={onButtonClick} />
     </div>
   );
 };